+++ example/all_signals.f90
@@ -1,7 +1,7 @@
 ! Forsynth: a multitracks stereo sound synthesis project
 ! License GPL-3.0-or-later
 ! Vincent Magnin
-! Last modifications: 2024-05-19
+! Last modifications: 2024-05-26
 
 ! Plays sequentially each type of available signal.
 program all_signals
@@ -12,18 +12,18 @@ program all_signals
                      & add_sawtooth_wave, add_triangle_wave, &
                      & add_karplus_strong, add_karplus_strong_stretched, &
                      & add_noise, add_weierstrass
-    use envelopes, only: attack, decay
+    use envelopes, only: ADSR_envelope
 
     implicit none
     type(WAV_file) :: demo
+    type(ADSR_envelope) :: env
     real(wp) :: t, Dt
     real(wp) :: f_A
 
     print *, "**** Demo of the available signals ****"
     call demo%create_WAV_file('all_signals.wav', tracks=1, duration=30._wp)
 
-    attack = 30.0_wp
-    decay  = 20.0_wp
+    call env%new(A=30._wp, D=20._wp, S=80._wp, R=30._wp)
 
     ! Notes frequencies:
     f_A = fr("A3")             ! A 220 Hz
@@ -32,22 +32,22 @@ program all_signals
 
     t = 0.0_wp
     print *, "Sinusoidal signal"
-    call add_sine_wave(demo%tape_recorder, track=1, t1=t, t2=t+Dt, f=f_A, Amp=1.0_wp)
+    call add_sine_wave(demo%tape_recorder, track=1, t1=t, t2=t+Dt, f=f_A, Amp=1.0_wp, envelope=env)
     print *, "Square wave"
-    call add_square_wave(demo%tape_recorder, track=1, t1=t+Dt, t2=t+2*Dt, f=f_A, Amp=1.0_wp)
+    call add_square_wave(demo%tape_recorder, track=1, t1=t+Dt, t2=t+2*Dt, f=f_A, Amp=1.0_wp, envelope=env)
     print *, "Sawtooth wave"
-    call add_sawtooth_wave(demo%tape_recorder, track=1, t1=t+2*Dt, t2=t+3*Dt, f=f_A, Amp=1.0_wp)
+    call add_sawtooth_wave(demo%tape_recorder, track=1, t1=t+2*Dt, t2=t+3*Dt, f=f_A, Amp=1.0_wp, envelope=env)
     print *, "Triangle wave"
-    call add_triangle_wave(demo%tape_recorder, track=1, t1=t+3*Dt, t2=t+4*Dt, f=f_A, Amp=1.0_wp)
+    call add_triangle_wave(demo%tape_recorder, track=1, t1=t+3*Dt, t2=t+4*Dt, f=f_A, Amp=1.0_wp, envelope=env)
     print *, "Summing the four signals together"
-    call add_sine_wave(demo%tape_recorder,     track=1, t1=t+4*Dt, t2=t+5*Dt, f=f_A, Amp=0.5_wp)
-    call add_square_wave(demo%tape_recorder,   track=1, t1=t+4*Dt, t2=t+5*Dt, f=f_A, Amp=0.5_wp)
-    call add_sawtooth_wave(demo%tape_recorder, track=1, t1=t+4*Dt, t2=t+5*Dt, f=f_A, Amp=0.5_wp)
-    call add_triangle_wave(demo%tape_recorder, track=1, t1=t+4*Dt, t2=t+5*Dt, f=f_A, Amp=0.5_wp)
+    call add_sine_wave(demo%tape_recorder,     track=1, t1=t+4*Dt, t2=t+5*Dt, f=f_A, Amp=0.5_wp, envelope=env)
+    call add_square_wave(demo%tape_recorder,   track=1, t1=t+4*Dt, t2=t+5*Dt, f=f_A, Amp=0.5_wp, envelope=env)
+    call add_sawtooth_wave(demo%tape_recorder, track=1, t1=t+4*Dt, t2=t+5*Dt, f=f_A, Amp=0.5_wp, envelope=env)
+    call add_triangle_wave(demo%tape_recorder, track=1, t1=t+4*Dt, t2=t+5*Dt, f=f_A, Amp=0.5_wp, envelope=env)
     print *, "Noise"
     call add_noise(demo%tape_recorder, track=1, t1=t+5*Dt, t2=t+6*Dt, Amp=1.0_wp)
     print *, "Weierstrass"
-    call add_weierstrass(demo%tape_recorder, track=1, t1=t+6*Dt, t2=t+7*Dt, f=f_A, Amp=1.0_wp)
+    call add_weierstrass(demo%tape_recorder, track=1, t1=t+6*Dt, t2=t+7*Dt, f=f_A, Amp=1.0_wp, envelope=env)
     print *, "Karplus Strong"
     call add_karplus_strong(demo%tape_recorder, track=1, t1=t+7*Dt, t2=t+8*Dt, f=f_A, Amp=1.0_wp)
     print *, "Karplus Strong stretched"