+++ example/demo_effects.f90
@@ -1,7 +1,7 @@
 ! Forsynth: a multitracks stereo sound synthesis project
 ! License GPL-3.0-or-later
 ! Vincent Magnin
-! Last modifications: 2024-05-19
+! Last modifications: 2024-05-26
 
 ! All available audio effects are applied sequentially on a chord sequence.
 program demo_effects
@@ -11,28 +11,28 @@ program demo_effects
     use music, only: fr, add_chord
     use audio_effects, only: apply_fuzz_effect, apply_tremolo_effect, &
                            & apply_autopan_effect, apply_delay_effect
-    use envelopes, only: attack, decay
+    use envelopes, only: ADSR_envelope
 
     implicit none
     type(WAV_file) :: demo
+    type(ADSR_envelope) :: env
     integer  :: i
     real(wp) :: t, Dt
 
     print *, "**** Demo of the audio effects ****"
     call demo%create_WAV_file('demo_effects.wav', tracks=1, duration=120._wp)
 
-    attack = 10.0_wp
-    decay  = 40.0_wp
+    call env%new(A=10._wp, D=40._wp, S=80._wp, R=30._wp)
 
     ! Notes duration in seconds:
     Dt = 1.5_wp
 
     print *, "Track 1: repeating G D F C chords..."
     t = 0.0_wp
-    call add_chord(demo%tape_recorder, track=1, t1=t,      t2=t+Dt,   f=fr("G3"), Amp=1.0_wp, chord=MAJOR_CHORD)
-    call add_chord(demo%tape_recorder, track=1, t1=t+Dt,   t2=t+2*Dt, f=fr("D3"), Amp=1.0_wp, chord=MAJOR_CHORD)
-    call add_chord(demo%tape_recorder, track=1, t1=t+2*Dt, t2=t+3*Dt, f=fr("F3"), Amp=1.0_wp, chord=MAJOR_CHORD)
-    call add_chord(demo%tape_recorder, track=1, t1=t+3*Dt, t2=t+4*Dt, f=fr("C3"), Amp=1.0_wp, chord=MAJOR_CHORD)
+    call add_chord(demo%tape_recorder, track=1, t1=t,      t2=t+Dt,   f=fr("G3"), Amp=1.0_wp, chord=MAJOR_CHORD, envelope=env)
+    call add_chord(demo%tape_recorder, track=1, t1=t+Dt,   t2=t+2*Dt, f=fr("D3"), Amp=1.0_wp, chord=MAJOR_CHORD, envelope=env)
+    call add_chord(demo%tape_recorder, track=1, t1=t+2*Dt, t2=t+3*Dt, f=fr("F3"), Amp=1.0_wp, chord=MAJOR_CHORD, envelope=env)
+    call add_chord(demo%tape_recorder, track=1, t1=t+3*Dt, t2=t+4*Dt, f=fr("C3"), Amp=1.0_wp, chord=MAJOR_CHORD, envelope=env)
     ! Repeat those four chords until the end of the track:
     do i = 1, 19
         call demo%copy_section(from_track=1, to_track=1, t1=t, t2=t+4*Dt, t3=4*Dt*i)